// backend/src/types/user.ts
export interface User{
    _id?: string;
    email?: string;
    password?: string
    role?: string;
    firstName?: string;
    lastName?: string;
    gender?: string;
    dateOfBirth?: string;
    employeeNumber?: string;
    designation?: string;
    department?: string;
    dateOfJoining?: string;
    address?: string;
    mobile?: string;
    alternateMobile?: string;
    bank?: string;
    bankAcNumber?: string;
    ifscCode?: string;
    salaryInfo?: string;
    skills?: string[];
}